refactor(app): make base example app use the project library
closes #4
