+++ app/build.gradle
@@ -68,4 +68,7 @@ dependencies {
     implementation "androidx.media3:media3-exoplayer:$mediaVersion"
     implementation "androidx.media3:media3-ui:$mediaVersion"
     implementation "androidx.media3:media3-exoplayer-dash:$mediaVersion"
+
+    // our library
+    implementation project(':androidenhancedvideoplayer')
 }
