+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Android Enhanced Video Player"
 include ':app'
+include ':androidenhancedvideoplayer'
